+++ Scripts/mcpTool.js
@@ -60,7 +60,7 @@ class MCPTool extends Tool {
                 name: this.name,
                 arguments: args,
                 _meta: {
-                    "io.modelcontextprotocol/protocolVersion": "2026-07-28",
+                    "io.modelcontextprotocol/protocolVersion": mcpAdapter.PROTOCOL_VERSION,
                     "io.modelcontextprotocol/clientInfo": {
                         name: nova.extension.name,
                         version: nova.extension.version